+++ src/test/resources/calculateStub/calculationDataSet15_Request.json
@@ -22,7 +22,8 @@
         "changeInLifetimeAllowancePercentageInformedFlag": true,
         "increaseInLTACharge": true,
         "newLTACharge": false,
-        "multipleBenefitCrystallisationEventFlag": false
+        "multipleBenefitCrystallisationEventFlag": true,
+        "otherSchemeNotification": true
       }
     },
     "annualAllowance": {
@@ -63,7 +64,6 @@
           "income": {
             "incomeAboveThreshold" : true,
             "adjustedIncome" : 325300
-
           },
           "totalIncome": 60000,
           "incomeSubJourney" : {
@@ -170,7 +170,7 @@
           },
           "income" : {
             "incomeAboveThreshold" : true,
-            "adjustedIncome" : 331450
+            "adjustedIncome" : 1450
           },
           "totalIncome": 300000,
           "chargePaidByMember": 800,
@@ -236,7 +236,7 @@
           },
           "income" : {
             "incomeAboveThreshold" : true,
-            "adjustedIncome" : 324000
+            "adjustedIncome" : 4000
           },
 
           "totalIncome": 300000,
@@ -272,7 +272,7 @@
           },
           "income" : {
             "incomeAboveThreshold" : true,
-            "adjustedIncome" : 324000
+            "adjustedIncome" : 4000
           },
           "totalIncome": 300000,
           "chargePaidByMember": 800,
@@ -336,8 +336,8 @@
     },
     "totalAmounts": {
       "outDatesCompensation": 3200,
-      "inDatesDebit": 43500,
-      "inDatesCredit": 800
+      "inDatesDebit": 0,
+      "inDatesCredit": 3200
     },
     "outDates": [
       {
@@ -432,13 +432,13 @@
     "inDates": [
       {
         "period": "2020",
-        "memberCredit": 0,
+        "memberCredit": 800,
         "schemeCredit": 0,
-        "debit": 12700,
+        "debit": 0,
         "chargePaidByMember": 800,
         "chargePaidBySchemes": 0,
-        "revisedChargableAmountBeforeTaxRate": 30000,
-        "revisedChargableAmountAfterTaxRate": 13500,
+        "revisedChargableAmountBeforeTaxRate": 0,
+        "revisedChargableAmountAfterTaxRate": 0,
         "unusedAnnualAllowance": 0,
         "taxYearSchemes": [
           {
@@ -478,13 +478,13 @@
       },
       {
         "period": "2022",
-        "memberCredit": 0,
+        "memberCredit": 800,
         "schemeCredit": 0,
-        "debit": 15400,
+        "debit": 0,
         "chargePaidByMember": 800,
         "chargePaidBySchemes": 0,
-        "revisedChargableAmountBeforeTaxRate": 36000,
-        "revisedChargableAmountAfterTaxRate": 16200,
+        "revisedChargableAmountBeforeTaxRate": 0,
+        "revisedChargableAmountAfterTaxRate": 0,
         "unusedAnnualAllowance": 0,
         "taxYearSchemes": [
           {
@@ -501,13 +501,13 @@
       },
       {
         "period": "2023",
-        "memberCredit": 0,
+        "memberCredit": 800,
         "schemeCredit": 0,
-        "debit": 15400,
+        "debit": 0,
         "chargePaidByMember": 800,
         "chargePaidBySchemes": 0,
-        "revisedChargableAmountBeforeTaxRate": 36000,
-        "revisedChargableAmountAfterTaxRate": 16200,
+        "revisedChargableAmountBeforeTaxRate": 0,
+        "revisedChargableAmountAfterTaxRate": 0,
         "unusedAnnualAllowance": 0,
         "taxYearSchemes": [
           {
