+++ src/test/resources/calculateStub/calculationDataSet4_Request.json
@@ -388,7 +388,7 @@
         "chargePaidByMember": 1200,
         "chargePaidBySchemes": 0,
         "revisedChargableAmountBeforeTaxRate": 283000,
-        "revisedChargableAmountAfterTaxRate": 122850,
+        "revisedChargableAmountAfterTaxRate": 122643,
         "unusedAnnualAllowance": 0,
         "taxYearSchemes": [
           {
