+++ src/test/resources/calculateStub/calculationDataSet5_Request.json
@@ -387,7 +387,7 @@
         "chargePaidByMember": 1200,
         "chargePaidBySchemes": 0,
         "revisedChargableAmountBeforeTaxRate": 283000,
-        "revisedChargableAmountAfterTaxRate": 122850,
+        "revisedChargableAmountAfterTaxRate": 122643,
         "unusedAnnualAllowance": 0,
         "taxYearSchemes": [
           {
