+++ src/test/resources/calculateStub/calculationDataSet6_Request.json
@@ -366,7 +366,7 @@
         "chargePaidByMember": 1200,
         "chargePaidBySchemes": 0,
         "revisedChargableAmountBeforeTaxRate": 283000,
-        "revisedChargableAmountAfterTaxRate": 122850,
+        "revisedChargableAmountAfterTaxRate": 122643,
         "unusedAnnualAllowance": 0,
         "taxYearSchemes": [
           {
