+++ src/test/scala/uk/gov/hmrc/test/ui/specs/SubmissionJourney13.scala
@@ -22,8 +22,13 @@ import uk.gov.hmrc.test.ui.pages._
 import uk.gov.hmrc.test.ui.specs.tags.ZapTests
 import util.CalculationDataUtil
 
+import scala.collection.mutable.ArrayBuffer
+
 class SubmissionJourney13 extends BaseSpec with BeforeAndAfter {
   var uniqueTaxSchemes: Map[String, String] = Map(("Scheme 1" -> "00348916RU"))
+  var debitYears: ArrayBuffer[Int]          = ArrayBuffer(
+    2020
+  )
 
   Feature("PRA Submission Journey 3, Business scenario AA journeys") {
 
@@ -60,6 +65,21 @@ class SubmissionJourney13 extends BaseSpec with BeforeAndAfter {
 
       When("I verify InternationalAddressPage, Enter Address information and continue")
       TheirInternationalAddressPage.verifyPageEnterAddressAndContinue()
+
+      debitYears.foreach { element =>
+        When("I verify WhoWillPayPage, select pension scheme and click continue button")
+        WhoWillPayPage.verifyPageSelectPensionSchemeAndContinue()
+
+        When("I verify WhichPensionSchemeWillPayPage, select public pension scheme and click continue button")
+        WhichPensionSchemeWillPayPage.verifyPageSelectPSAndContinue("Scheme 1")
+
+        When("I verify Valid Election for Scheme to pay Page and select no and click continue")
+        AskedPensionSchemeToPayTaxCharge.verifyPageSelectNoAndContinue()
+
+        When("I verify SchemeElectionConsent Page, click Agree and Continue")
+        SchemeElectionConsentPage.verifySchemeElectionConsentPageClickAgreeAndContinue()
+
+      }
 
       When("I verify AlternativeNamePage, select No and continue")
       AlternativeNamePage.verifyPageSelectYesAndContinue()
