+++ src/test/scala/uk/gov/hmrc/test/ui/specs/SubmissionJourney8.scala
@@ -22,11 +22,17 @@ import uk.gov.hmrc.test.ui.pages._
 import uk.gov.hmrc.test.ui.specs.tags.ZapTests
 import util.CalculationDataUtil
 
+import scala.collection.mutable.ArrayBuffer
+
 class SubmissionJourney8 extends BaseSpec with BeforeAndAfter {
   var uniqueTaxSchemes: Map[String, String] = Map(
     /** add scheme details from the test json to below map * */
     ("Scheme 1" -> "003 48916ed"),
     ("Scheme 4" -> "00348916 uC")
+  )
+
+  var debitYears: ArrayBuffer[Int] = ArrayBuffer(
+    2020
   )
 
   Feature("Business scenario AA journeys") {
@@ -65,6 +71,21 @@ class SubmissionJourney8 extends BaseSpec with BeforeAndAfter {
 
       When("I verify TheirUKAddressPage, Enter Address information and continue")
       TheirUKAddressPage.verifyPageEnterAddressAndContinue()
+
+      debitYears.foreach { element =>
+        When("I verify WhoWillPayPage, select pension scheme and click continue button")
+        WhoWillPayPage.verifyPageSelectPensionSchemeAndContinue()
+
+        When("I verify WhichPensionSchemeWillPayPage, select public pension scheme and click continue button")
+        WhichPensionSchemeWillPayPage.verifyPageSelectPSAndContinue("Scheme 1")
+
+        When("I verify Valid Election for Scheme to pay Page and select no and click continue")
+        AskedPensionSchemeToPayTaxCharge.verifyPageSelectNoAndContinue()
+
+        When("I verify SchemeElectionConsent Page, click Agree and Continue")
+        SchemeElectionConsentPage.verifySchemeElectionConsentPageClickAgreeAndContinue()
+
+      }
 
       When("I verify AlternativeNamePage, select No and continue")
       AlternativeNamePage.verifyPageSelectNoAndContinue()
